pyproject.toml
[project]
name = "python-sql"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.4.0",
]
